# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Get a list of ingredients",
                "operationId": "getIngredients",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IngredientResponseSchema"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Token missing or invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized. Token missing or invalid."
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "error_type": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Create a new ingredient",
                "operationId": "createIngredient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IngredientRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingredient created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Request. Validation errors.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unprocessable Request."
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "error_type": {
                                            "type": "string",
                                            "example": "Unprocessable Content"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Token missing or invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized. Token missing or invalid."
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "error_type": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "IngredientRequestSchema": {
                "title": "Ingredient Request Schema",
                "description": "Schema for creating an ingredient",
                "required": [
                    "name",
                    "quantity"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "IngredientResponseSchema": {
                "title": "Ingredient Response Schema",
                "description": "Schema for an ingredient response",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Tomato",
                    "quantity": 20,
                    "updated_at": "2023-08-28T00:54:54.000000Z",
                    "created_at": "2023-08-28T00:54:54.000000Z",
                    "id": 1
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}